refactor: PromotionApi의 id 타입을 Int로 변경
Swagger에 명시되어 있는대로 id타입을 Long에서 Int로 변경

issue: #46

+++ domain/src/main/java/buy/coke/zet/domain/entitiy/promotion/StorePromotionResponseEntity.kt
@@ -6,7 +6,7 @@ data class StorePromotionResponseEntity(
 )
 
 data class PromotionItemEntity(
-    val productId: Long,
+    val productId: Int,
     val productName: String,
     val size: String,
     val promotionTypeLabel: String,
